fix for stm32f765/767

+++ rcan_timing.h
@@ -6,9 +6,9 @@
 
 typedef struct {
     uint16_t bit_rate_prescaler;                        /// [1, 1024]
-    uint8_t bit_segment_1;                              /// [1, 16]
-    uint8_t bit_segment_2;                              /// [1, 8]
-    uint8_t max_resynchronization_jump_width;           /// [1, 4] (recommended value is 1)
+    uint32_t bit_segment_1;                              /// [1, 16]
+    uint32_t bit_segment_2;                              /// [1, 8]
+    uint32_t max_resynchronization_jump_width;           /// [1, 4] (recommended value is 1)
 } rcan_timing;
